requirements.txt
streamlit==1.29.0
neo4j==5.15.0
groq==0.5.0
python-dotenv==1.0.0
